requirements.txt
fastapi
uvicorn[standard]
discord.py
python-dotenv
requests